refactor: rename Checkout component to CheckoutPage and update related imports; implement order creation and preference handling
